REV: PR #1631 review feedback — code cleanup across V&V'ed filters
Addresses the inline-comment batch from @JDuffeyBQ on PR #1631 plus
findings from a parallel adversarial sweep across the same source set.

Code cleanup:
- Strip meaningless `inline` keyword from free functions in anonymous /
  named namespaces of 4 test TUs (14 sites total — only 4 were originally
  flagged; the adversarial pass found 10 more in the `ToyFixtures`
  namespace of `ComputeFeatureNeighborCAxisMisalignmentsTest.cpp`).
- Drop "wrapped in an anonymous namespace so every symbol below has
  internal linkage" boilerplate from 5 test files (1 flagged + 4 sibling
  copies across the V&V cohort).
- Tighten divisor-variable scope from function-scope to outer-loop-scope
  in ComputeFeatureNeighborMisorientations.cpp and ComputeFeatureNeighbor-
  CAxisMisalignments.cpp; both lift the per-feature reset out of an
  inner-loop reassignment hazard.
- Convert ad-hoc `CAPTURE(errorMsg); REQUIRE(false);` failure-printing to
  Catch2's idiomatic `INFO() ; REQUIRE(actual == expected)` at two sites
  in BadDataNeighborOrientationCheckTest.cpp.
- Replace hand-rolled "arrays differ" loops with `!std::equal(a, b)` at
  two sites in ComputeIPFColorsTest.cpp + ComputeFaceIPFColoringTest.cpp.
- Drop a redundant `static_cast<uint32>(ebsdlib::CrystalStructure::
  UnknownCrystalStructure)` — the EbsdLib 3.0 enum value is already
  `uint32_t`.

Enum-default-fallthrough hardening:
- Make ColorKey choice → ebsdlib::ColorKeyKind conversion explicit at
  both consumers (ComputeIPFColorsFilter and ComputeFaceIPFColoringFilter)
  — enumerate cases 0/1/2 by hand; out-of-range now returns an explicit
  error instead of silently defaulting to TSL. JDuffey: *"A value out of
  range of an enum is an error."*
- Same treatment for HexConvention choice in WritePoleFigureFilter — the
  previous bare `static_cast<HexConvention>(choice)` is replaced with a
  validated switch.
- Algorithm-side: silent `default: break;` on a CrystalStructure switch
  in WritePoleFigure.cpp's per-phase intensity-pole-figure loop now emits
  a Type::Warning naming the phase index + unknown crystal-structure
  value, so a phase with an out-of-spec crystal structure is visible in
  the run log instead of silently producing no pole figures.

+++ src/Plugins/OrientationAnalysis/src/OrientationAnalysis/Filters/Algorithms/BadDataNeighborOrientationCheck.cpp
@@ -82,7 +82,7 @@ Result<> BadDataNeighborOrientationCheck::operator()()
   const usize numOrientationOps = orientationOps.size();
   for(usize i = 0; i < crystalStructures.getSize(); ++i)
   {
-    if(crystalStructures[i] >= numOrientationOps && crystalStructures[i] != static_cast<uint32>(ebsdlib::CrystalStructure::UnknownCrystalStructure))
+    if(crystalStructures[i] >= numOrientationOps && crystalStructures[i] != ebsdlib::CrystalStructure::UnknownCrystalStructure)
     {
       return MakeErrorResult(
           -54901, fmt::format("Crystal structure at ensemble index {} has value {}, which is not a valid Laue-group index. Valid range is [0, {}).", i, crystalStructures[i], numOrientationOps));
